pyproject.toml
[tool.poetry]
name = "plus-db-agent"
version = "0.1.0"
description = ""
authors = ["Pedro Gustavo Santana <pedrogustavosantan97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
invoke = "^2.2.0"
pydantic = "^2.7.2"
aerich = "^0.7.2"
bcrypt = "^4.1.3"
passlib = "^1.7.4"
PyJWT = "^2.8.0"
pytest = "^8.2.1"
pytest-asyncio = "^0.23.7"
tortoise-orm = "^0.21.2"
email-validator = "^2.1.1"
python-dotenv = "^1.0.1"
fastapi-filter = "^2.0.0"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pre-commit = "^3.7.1"


[tool.aerich]
tortoise_orm = "database_agent.config.DATABASE_CONFIG"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
